+++ src/idunno.Bluesky/Notifications/BlueskyAgent.cs
@@ -319,6 +319,28 @@ public partial class BlueskyAgent
             throw new AuthenticationRequiredException();
         }
 
+        return await SetNotificationPreferences(preferences, cancellationToken).ConfigureAwait(false);
+    }
+
+    /// <summary>
+    /// Sets the current user's notification preferences.
+    /// </summary>
+    /// <param name="preferences">The notification preferences to set.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="preferences"/> is <see langword="null"/>.</exception>
+    /// <exception cref="AuthenticationRequiredException">Thrown when the agent is not authenticated.</exception>
+    public async Task<AtProtoHttpResult<Notifications.Preferences>> PutNotificationPreferences(
+        Notifications.Preferences preferences,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(preferences);
+
+        if (!IsAuthenticated)
+        {
+            throw new AuthenticationRequiredException();
+        }
+
         return await BlueskyServer.PutNotificationPreferences(
             preferences: preferences,
             service: Service,
@@ -328,4 +350,6 @@ public partial class BlueskyAgent
             loggerFactory: LoggerFactory,
             cancellationToken: cancellationToken).ConfigureAwait(false);
     }
+
 }
+