+++ src/idunno.Bluesky/Notifications/Model/PutPreferencesV2Request.cs
@@ -6,7 +6,9 @@ using idunno.Bluesky.Notifications.PreferenceTypes;
 namespace idunno.Bluesky.Notifications.Model;
 
 internal sealed record PutPreferencesV2Request(
+#pragma warning disable CS0618 // Type or member is obsolete, kept to allow deserialization of old data
     ChatPreference Chat,
+#pragma warning restore CS0618 // Type or member is obsolete
     FilterablePreference Follow,
     FilterablePreference LikeViaRepost,
     FilterablePreference Mention,